Fixes #52
No extractor worked since the jar-with-dependencies were not generated (due to a profile activation change). So I used the oldest profile activation strategy (which always work): creating a marker file in the projects which requires that activation.
I *will* change that for something cleaner later on. But for now, I want to test fast that url adding work in every case.